refactor(1013): extract OrgTicketManager facade (SC-001, Cat B position 46)
Extract OrgTicketManager (~475 LOC, 6 public methods + 19 private helpers)
from MistHelper.py to src/org/org_ticket_manager.py under the SC-001 facade
pattern. Live-global access goes through lazy importlib.import_module("MistHelper")
inside each method to preserve original call semantics.

- src/org/org_ticket_manager.py: new module with full class body
- MistHelper.py: add re-export alias with F401/pylint breadcrumb; class body
  replaced with single-line breadcrumb comment
- pyproject.toml: add coverage omit entry (alphabetical)

Local gates: black clean, ruff clean, py_compile OK, --test baseline
66 covered / 131 skipped / 0 failed (unchanged).
