+++ MistHelper.py
@@ -220,6 +220,9 @@ from src.inventory.org_device_inventory_summary_facade import (  # pylint: disab
     OrgDeviceInventorySummary,  # noqa: F401  # Cat B (1013 SC-001 position 29) -- re-export for MistHelper.OrgDeviceInventorySummary callers
 )
 from src.org.org_config_migration_manager import OrgConfigMigrationManager  # Cat B (1013 SC-001 position 5)
+from src.org.org_ticket_manager import (  # pylint: disable=unused-import
+    OrgTicketManager,  # noqa: F401  # Cat B (1013 SC-001 position 46) -- re-export for MistHelper.OrgTicketManager callers
+)
 from src.org_data_collector import OrgDataCollector  # Import org-level data collection orchestrator
 from src.refactors.anomaly_metrics_discovery import (  # pylint: disable=unused-import
     AnomalyMetricsDiscovery,  # noqa: F401  # Cat B (1013 SC-001 pos 43) -- lazy access via mh.AnomalyMetricsDiscovery
@@ -7090,484 +7093,7 @@ class PromptUtils:  # General prompt helpers.
 # NOTE: DeviceUtils has been extracted to src/device/device_utils.py (issue #1013 SC-001 position 6)
 
 
-# ============================================================================
-# ORGANIZATION TICKET MANAGER CLASS
-# ============================================================================
-class OrgTicketManager:  # Support ticket operations.
-    """
-    Full lifecycle management for Juniper Mist support tickets.
-
-    Provides 6 public operations (list, create, add comment, update, view, export) that
-    cover reading, creating, and modifying support tickets via the Mist API.
-    Attachment support is integrated into add_comment via multipart upload.
-    """
-
-    TICKET_TYPES = ["question", "problem", "incident", "feature_request"]  # Valid Mist ticket type values
-
-    # ------------------------------------------------------------------
-    # Public entry points (6 operations -- cohesive ticket lifecycle)
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def list_tickets() -> None:  # List org support tickets.
-        """Menu 188: Export all organization support tickets to CSV/SQLite."""
-        logging.info("Menu 188: Starting organization ticket list export")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.list_tickets()")  # Debug trace
-        try:
-            APIDataFetcher(  # Delegate to standard fetch-export pipeline
-                title="Organization Support Tickets:",  # User-facing header
-                api_call=mistapi.api.v1.orgs.tickets.listOrgTickets,  # SDK function for ticket listing
-                filename="OrgTickets.csv",  # Output filename in data/ directory
-                sort_key="created_at",  # Sort tickets by creation timestamp
-                duration="365d",  # Look back 1 year (SDK defaults to 1d which misses older tickets)
-            ).execute()  # Run the full fetch-flatten-export workflow
-            logging.info("Completed org ticket list export")  # Log success
-            logging.debug("EXIT: OrgTicketManager.list_tickets - success")  # Debug trace
-        except Exception as error:  # Catch API or export failures
-            logging.error("Failed to export org tickets: %s", error)  # Log error with context
-            logging.debug("EXIT: OrgTicketManager.list_tickets - error")  # Debug trace
-            raise  # Re-raise so caller sees the failure
-
-    @staticmethod
-    def create_ticket() -> None:  # Create a support ticket.
-        """Menu 189: Create a new support ticket in the organization."""
-        logging.info("Menu 189: Starting support ticket creation")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.create_ticket()")  # Debug trace
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve org from cache or user prompt
-        subject = OrgTicketManager._prompt_subject()  # Prompt user for ticket subject line
-        if not subject:  # User left subject blank -- abort
-            print("  Ticket creation cancelled -- subject is required.")  # Inform user
-            logging.info("Ticket creation cancelled: blank subject")  # Log cancellation
-            return  # Early exit
-        ticket_type = OrgTicketManager._prompt_ticket_type()  # Prompt user to select ticket type
-        comment = InputUtils.safe_input(  # Prompt for initial ticket description (optional)
-            "  Enter initial comment/description: ",
-            default_value="",
-            allow_empty=True,
-            context="create_ticket_comment",
-        )
-        body = {"subject": subject, "type": ticket_type}  # Build required API request fields
-        if comment:  # Include comment only if user provided one
-            body["comment"] = comment  # Add optional comment to request body
-        OrgTicketManager._submit_create_ticket(org_id, body, subject, ticket_type)  # API + report
-
-    @staticmethod
-    def add_comment() -> None:  # Add a comment to a ticket.
-        """Menu 190: Add a comment (with optional attachment) to an existing ticket."""
-        logging.info("Menu 190: Starting add comment to ticket")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.add_comment()")  # Debug trace
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve org from cache or user prompt
-        ticket_id = OrgTicketManager._select_ticket(org_id)  # Show ticket list for user selection
-        if not ticket_id:  # User cancelled selection -- abort
-            print("  Operation cancelled -- no ticket selected.")  # Inform user
-            logging.info("Add comment cancelled: no ticket selected")  # Log the cancellation
-            return  # Early exit without adding comment
-        comment_text, file_path = OrgTicketManager._prompt_comment_and_file()  # Gather inputs together
-        if not comment_text and not file_path:  # Neither comment nor file provided -- abort
-            print("  Operation cancelled -- provide a comment or file.")  # Inform user
-            logging.info("Add comment cancelled: no comment or file provided")  # Log cancellation
-            return  # Early exit
-        OrgTicketManager._submit_comment(  # Submit comment to API
-            org_id,
-            ticket_id,
-            comment_text,
-            file_path,  # Pass all user-provided values
-        )
-
-    @staticmethod
-    def update_ticket() -> None:  # Update a support ticket.
-        """Menu 191: Update fields on an existing support ticket."""
-        logging.info("Menu 191: Starting ticket update")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.update_ticket()")  # Debug trace
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve org from cache or user prompt
-        ticket_id = OrgTicketManager._select_ticket(org_id)  # Show ticket list for user selection
-        if not ticket_id:  # User cancelled selection -- abort
-            print("  Operation cancelled -- no ticket selected.")  # Inform user
-            logging.info("Ticket update cancelled: no ticket selected")  # Log the cancellation
-            return  # Early exit
-        body = OrgTicketManager._build_update_body()  # Collect changed fields from user
-        if not body:  # No fields were changed -- abort
-            print("  No changes specified -- update cancelled.")  # Inform user
-            logging.info("Ticket update cancelled: no fields changed")  # Log cancellation
-            return  # Early exit
-        OrgTicketManager._update_via_api(org_id, ticket_id, body)  # Send update + report results
-
-    # ------------------------------------------------------------------
-    # Private helpers (max 5 per group)
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _prompt_subject() -> str:  # Prompt for ticket subject.
-        """Prompt user for ticket subject line."""
-        return InputUtils.safe_input(  # Use EOF-safe input wrapper
-            "  Enter ticket subject: ",  # Prompt text for ticket title
-            default_value="",  # No default -- user must provide subject
-            allow_empty=True,  # Allow blank to signal cancellation
-            context="create_ticket_subject",  # Context label for EOF logging
-        )
-
-    @staticmethod
-    def _prompt_ticket_type() -> str:  # Prompt for ticket type.
-        """Prompt user to select a ticket type from valid options."""
-        print("\n  Ticket types:")  # Section header for type selection
-        for index, ticket_type in enumerate(OrgTicketManager.TICKET_TYPES, 1):  # Number each type for selection
-            print(f"    {index}. {ticket_type}")  # Display numbered option
-        choice = InputUtils.safe_input(  # Prompt user to pick a type number
-            "  Select type [1]: ",  # Default to first option (question)
-            default_value="1",  # Default selection is 'question'
-            allow_empty=True,  # Allow enter for default
-            context="create_ticket_type",  # Context label for EOF logging
-        )
-        try:
-            index = int(choice) - 1  # Convert 1-based user input to 0-based index
-            if 0 <= index < len(OrgTicketManager.TICKET_TYPES):  # Validate index is within bounds
-                return OrgTicketManager.TICKET_TYPES[index]  # Return selected ticket type string
-        except ValueError:  # User entered non-numeric input
-            pass  # Fall through to default
-        return OrgTicketManager.TICKET_TYPES[0]  # Default to 'question' for invalid input
-
-    @staticmethod
-    def _prompt_ticket_id() -> str:  # Prompt for ticket id.
-        """Prompt user for ticket UUID."""
-        return InputUtils.safe_input(  # Use EOF-safe input wrapper
-            "  Enter ticket ID: ",  # Prompt text for ticket UUID
-            default_value="",  # No default -- user must provide ID
-            allow_empty=True,  # Allow blank to signal cancellation
-            context="ticket_id_prompt",  # Context label for EOF logging
-        )
-
-    @staticmethod
-    def _print_ticket_created_summary(ticket_data: dict, subject: str, ticket_type: str) -> None:
-        """Print + log the newly created ticket summary."""
-        ticket_id = ticket_data.get("id", "unknown")  # Get new ticket UUID from response
-        logging.debug("Ticket created: id=%s, status=%s", ticket_id, ticket_data.get("status"))  # Log result
-        print("\n  Ticket created successfully!")  # Confirm to user
-        print(f"  ID:      {ticket_id}")  # Display ticket ID for reference
-        print(f"  Subject: {subject}")  # Echo subject back to user
-        print(f"  Type:    {ticket_type}")  # Echo type back to user
-        print(f"  Status:  {ticket_data.get('status', 'open')}")  # Show initial status
-        logging.info("Menu 189: Ticket creation complete, id=%s", ticket_id)  # Log success
-
-    @staticmethod
-    def _submit_create_ticket(org_id: str, body: dict, subject: str, ticket_type: str) -> None:
-        """Send createOrgTicket API + print summary (or print + raise on error)."""
-        logging.info("Creating ticket '%s' (type=%s) in org %s", subject, ticket_type, org_id)
-        try:
-            response = mistapi.api.v1.orgs.tickets.createOrgTicket(apisession, org_id, body)
-            OrgTicketManager._print_ticket_created_summary(getattr(response, "data", {}), subject, ticket_type)
-        except Exception as error:  # Catch API errors during ticket creation
-            logging.error("Failed to create ticket: %s", error)  # Log error with context
-            print(f"\n  Error creating ticket: {error}")  # Show error to user
-            raise  # Re-raise for upstream error handling
-
-    @staticmethod
-    def _build_update_body() -> dict[str, str]:  # Build ticket update body.
-        """Collect optional update fields (subject, status, type) from user prompts."""
-        body: dict[str, str] = {}  # Accumulate changed fields in a dict
-        fields = (  # (api_key, prompt_text, eof_context) tuples for each updatable field
-            ("subject", "  New subject (leave blank to skip): ", "update_ticket_subject"),
-            ("status", "  New status [open/closed] (leave blank to skip): ", "update_ticket_status"),
-            (
-                "type",
-                "  New type [question/problem/incident/feature_request] (leave blank to skip): ",
-                "update_ticket_type",
-            ),
-        )
-        for api_key, prompt, ctx in fields:  # Prompt for each updatable field in turn
-            value = InputUtils.safe_input(  # EOF-safe prompt for this field
-                prompt,
-                default_value="",
-                allow_empty=True,
-                context=ctx,
-            )
-            if value:  # Only include field if user provided a value
-                body[api_key] = value  # Add user-supplied value to update body
-        return body  # Return dict of fields to update (may be empty)
-
-    @staticmethod
-    def _update_via_api(org_id: str, ticket_id: str, body: dict[str, str]) -> None:
-        """Send updateOrgTicket API call and print + log results."""
-        logging.info("Updating ticket %s with fields: %s", ticket_id, list(body.keys()))  # Log before API call
-        try:
-            response = mistapi.api.v1.orgs.tickets.updateOrgTicket(  # Call Mist API to update ticket
-                apisession,
-                org_id,
-                ticket_id,
-                body,  # Pass session, org, ticket ID, body
-            )
-            logging.debug("Ticket updated: %s", getattr(response, "data", {}))  # Log full response
-            print(f"\n  Ticket {ticket_id} updated successfully!")  # Confirm to user
-            for field, value in body.items():  # Show each changed field to user
-                print(f"  {field}: {value}")  # Display field name and new value
-            logging.info("Menu 191: Ticket update complete for %s", ticket_id)  # Log success
-        except Exception as error:  # Catch API errors during ticket update
-            logging.error("Failed to update ticket %s: %s", ticket_id, error)  # Log error with context
-            print(f"\n  Error updating ticket: {error}")  # Show error to user
-            raise  # Re-raise for upstream error handling
-
-    @staticmethod
-    def _prompt_comment_and_file() -> tuple[str, str]:
-        """Prompt for comment text and optional attachment path."""
-        comment_text = InputUtils.safe_input(  # Prompt user for comment body text
-            "  Enter comment text: ",
-            default_value="",
-            allow_empty=True,
-            context="add_ticket_comment",
-        )
-        file_path = InputUtils.safe_input(  # Prompt for optional file attachment path
-            "  Attach a file? Enter path (leave blank to skip): ",
-            default_value="",
-            allow_empty=True,
-            context="add_ticket_attachment",
-        )
-        return comment_text, file_path  # Tuple of (text, file_path)
-
-    @staticmethod
-    def _submit_comment(org_id: str, ticket_id: str, comment_text: str, file_path: str) -> None:
-        """Submit comment with optional file attachment to ticket."""
-        has_file = bool(file_path and os.path.isfile(file_path))  # Check if valid file was specified
-
-        if has_file:  # Use multipart upload API when file is attached
-            logging.info("Adding comment with attachment to ticket %s", ticket_id)  # Log before API call
-            mistapi.api.v1.orgs.tickets.addOrgTicketCommentFile(  # Multipart comment+file API
-                apisession,
-                org_id,
-                ticket_id,  # Session, org, and ticket identifiers
-                comment=comment_text or None,  # Comment text (None if empty)
-                file=file_path,  # Path to file for upload
-            )
-            logging.debug("Comment with file submitted to ticket %s", ticket_id)  # Log after API call
-            print(f"\n  Comment with attachment added to ticket {ticket_id}")  # Confirm to user
-        elif file_path:  # User specified a path but file doesn't exist
-            logging.warning(  # Warn about missing file path
-                "File not found: %s -- adding comment without attachment", file_path
-            )
-            print(f"  Warning: File not found at '{file_path}' -- adding comment only.")  # Alert user
-            OrgTicketManager._submit_text_comment(org_id, ticket_id, comment_text)  # Fall back to text-only
-        else:  # No file specified -- text-only comment
-            OrgTicketManager._submit_text_comment(org_id, ticket_id, comment_text)  # Submit text comment
-
-    @staticmethod
-    def _submit_text_comment(org_id: str, ticket_id: str, comment_text: str) -> None:  # Submit a text-only comment.
-        """Submit a text-only comment to a ticket."""
-        logging.info("Adding text comment to ticket %s", ticket_id)  # Log before API call
-        body = {"comment": comment_text}  # Build comment request body
-        mistapi.api.v1.orgs.tickets.addOrgTicketComment(  # Call Mist API to add comment
-            apisession,
-            org_id,
-            ticket_id,
-            body,  # Session, org, ticket ID, and comment body
-        )
-        logging.debug("Text comment submitted to ticket %s", ticket_id)  # Log after API call
-        print(f"\n  Comment added to ticket {ticket_id}")  # Confirm to user
-
-    # ------------------------------------------------------------------
-    # Public entry points -- ticket viewing and export (Menu 192-193)
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def view_ticket() -> None:  # View a single ticket.
-        """Menu 192: View a single ticket with full comments and history."""
-        logging.info("Menu 192: Starting ticket detail viewer")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.view_ticket()")  # Debug trace
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve org from cache or user prompt
-
-        ticket_id = OrgTicketManager._select_ticket(org_id)  # Show ticket list for user selection
-        if not ticket_id:  # User cancelled selection -- abort
-            print("  Operation cancelled -- no ticket selected.")  # Inform user
-            logging.info("View ticket cancelled: no ticket selected")  # Log the cancellation
-            return  # Early exit
-
-        ticket_data = OrgTicketManager._fetch_ticket_detail(org_id, ticket_id)  # Fetch full ticket+comments
-        if not ticket_data:  # API returned empty or failed
-            print("  Could not retrieve ticket details.")  # Inform user of failure
-            return  # Early exit
-
-        OrgTicketManager._display_ticket_detail(ticket_data)  # Format and print to screen
-        logging.info("Menu 192: Ticket detail view complete for %s", ticket_id)  # Log success
-
-    @staticmethod
-    def export_ticket_details() -> None:  # Export ticket details to file.
-        """Menu 193: Export all tickets with full details and comments to CSV/SQLite."""
-        logging.info("Menu 193: Starting full ticket detail export")  # Log operation entry
-        logging.debug("ENTRY: OrgTicketManager.export_ticket_details()")  # Debug trace
-        org_id = ConfigUtils.get_cached_or_prompted_org_id()  # Resolve org from cache or user prompt
-        tickets = OrgTicketManager._fetch_all_ticket_summaries(org_id)  # API list summaries
-        if not tickets:  # No tickets found in the org
-            print("\n  No tickets found in this organization.")  # Inform user
-            return  # Nothing to export
-        all_details = OrgTicketManager._collect_ticket_details(org_id, tickets)  # Per-ticket detail fetch
-        if all_details:  # Export if we have any ticket details
-            logging.info("Exporting %d ticket details", len(all_details))  # Log before export
-            DataExporter.write_with_format_selection(  # Write to CSV/SQLite via standard pipeline
-                all_details,
-                "OrgTicketDetails.csv",
-                api_function_name="getOrgTicket",
-            )
-            logging.info("Menu 193: Full ticket detail export complete")  # Log success
-        else:  # No details were retrieved
-            print("\n  No ticket details could be retrieved.")  # Inform user
-
-    # ------------------------------------------------------------------
-    # Private helpers -- ticket selection and detail display
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _select_ticket(org_id: str) -> str:  # Prompt to select a ticket.
-        """List tickets and let user pick by index, or enter ID manually."""
-        tickets = OrgTicketManager._fetch_tickets_for_selection(org_id)  # Fetch + handle empty
-        if not tickets:  # No tickets or fetch error
-            return ""  # Signal cancellation to caller
-        OrgTicketManager._render_ticket_list_table(tickets)  # Display numbered ticket table
-        print(f"\n  Enter a number (1-{len(tickets)}) to select, or 'm' to enter ID manually.")
-        choice = InputUtils.safe_input(  # Prompt user for selection input
-            "  Selection: ",
-            default_value="",
-            allow_empty=True,
-            context="select_ticket",
-        )
-        if not choice:  # Blank input -- cancel
-            return ""  # Signal cancellation
-        if choice.lower() == "m":  # Manual ID entry path
-            return OrgTicketManager._prompt_ticket_id()  # Prompt for manual ticket UUID
-        return OrgTicketManager._resolve_ticket_choice(choice, tickets)  # Parse + validate index
-
-    @staticmethod
-    def _fetch_tickets_for_selection(org_id: str) -> list:
-        """Fetch ticket summaries for selection; print + return [] on error/empty."""
-        logging.info("Fetching ticket list for selection (org %s)", org_id)  # Log before API call
-        try:
-            response = mistapi.api.v1.orgs.tickets.listOrgTickets(  # Fetch ticket summaries
-                apisession,
-                org_id,
-                duration="365d",  # 1-year history window
-            )
-            tickets = getattr(response, "data", []) or []  # Extract ticket list
-            logging.debug("Retrieved %d tickets for selection", len(tickets))  # Log count
-        except Exception as error:  # Catch API failures
-            logging.error("Failed to fetch tickets for selection: %s", error)  # Log error
-            print(f"  Error fetching tickets: {error}")  # Show error to user
-            return []  # Signal failure with empty list
-        if not tickets:  # API returned no tickets
-            print("\n  No tickets found in this organization.")  # Inform user
-        return tickets  # Return list (possibly empty)
-
-    @staticmethod
-    def _render_ticket_list_table(tickets: list) -> None:
-        """Print a numbered table of ticket #/status/type/subject for selection."""
-        print("\n  Organization Support Tickets:")  # Section header
-        print(f"  {'#':<4} {'Status':<10} {'Type':<18} {'Subject'}")  # Column headers
-        print(f"  {'-' * 4} {'-' * 10} {'-' * 18} {'-' * 40}")  # Separator line
-        for index, ticket in enumerate(tickets, 1):  # Display numbered rows
-            status = ticket.get("status", "unknown")  # Ticket status field
-            ttype = ticket.get("type", "unknown")  # Ticket type field
-            subject = ticket.get("subject", "(no subject)")  # Ticket subject field
-            print(f"  {index:<4} {status:<10} {ttype:<18} {subject}")  # Print formatted row
-
-    @staticmethod
-    def _resolve_ticket_choice(choice: str, tickets: list) -> str:
-        """Parse numeric choice into ticket ID; print error + return '' on bad input."""
-        try:
-            idx = int(choice) - 1  # Convert 1-based to 0-based index
-        except ValueError:  # Non-numeric input that wasn't 'm'
-            print(f"  Invalid selection: {choice}")  # Inform user of bad input
-            return ""  # Signal cancellation
-        if not 0 <= idx < len(tickets):  # Index out of range
-            print(f"  Invalid selection: {choice}")  # Inform user of bad input
-            return ""  # Signal cancellation
-        selected_id = tickets[idx].get("id", "")  # Extract ticket ID
-        selected_subj = tickets[idx].get("subject", "(no subject)")  # Extract subject for confirmation
-        print(f"  Selected: {selected_subj}")  # Confirm selection to user
-        logging.info("User selected ticket %s (%s)", selected_id, selected_subj)  # Log selection
-        return selected_id  # Return chosen ticket ID
-
-    @staticmethod
-    def _fetch_ticket_detail(org_id: str, ticket_id: str) -> dict:  # Fetch one ticket detail.
-        """Fetch full ticket data including comments via getOrgTicket."""
-        logging.info("Fetching detail for ticket %s", ticket_id)  # Log before API call
-        try:
-            response = mistapi.api.v1.orgs.tickets.getOrgTicket(  # Call SDK for full ticket detail
-                apisession,
-                org_id,
-                ticket_id,
-                duration="365d",  # Look back 1 year for comment history
-            )
-            ticket_data = getattr(response, "data", {}) or {}  # Extract response data dict
-            logging.debug("Received ticket detail: %d fields", len(ticket_data))  # Log field count
-            return ticket_data  # Return the full ticket dict with comments
-        except Exception as error:  # Catch API failures
-            logging.error("Failed to fetch ticket %s: %s", ticket_id, error)  # Log error
-            print(f"  Error fetching ticket {ticket_id}: {error}")  # Show error to user
-            return {}  # Return empty dict to signal failure
-
-    @staticmethod
-    def _fetch_all_ticket_summaries(org_id: str) -> list:
-        """Fetch ticket-summary list via listOrgTickets, raise on failure."""
-        logging.info("Fetching ticket list for org %s", org_id)  # Log before API call
-        try:
-            response = mistapi.api.v1.orgs.tickets.listOrgTickets(  # Fetch all ticket summaries
-                apisession,
-                org_id,
-                duration="365d",  # 1-year window for ticket history
-            )
-            tickets = getattr(response, "data", []) or []  # Extract list from APIResponse
-            logging.debug("Found %d tickets to export with details", len(tickets))  # Log count
-            return tickets  # Return summary list to caller
-        except Exception as error:  # Catch API failures on ticket list
-            logging.error("Failed to fetch ticket list: %s", error)  # Log the failure
-            print(f"\n  Error fetching tickets: {error}")  # Show error to user
-            raise  # Re-raise for upstream handling
-
-    @staticmethod
-    def _collect_ticket_details(org_id: str, tickets: list) -> list:
-        """For each summary in tickets, fetch + flatten its full detail. Returns list of flat dicts."""
-        all_details = []  # Accumulate flattened ticket+comment records
-        print(f"\n  Fetching details for {len(tickets)} tickets...")  # Progress indicator
-        for index, ticket in enumerate(tickets, 1):  # Iterate each ticket summary
-            tid = ticket.get("id", "")  # Extract ticket ID from summary
-            if not tid:  # Skip tickets without valid IDs
-                continue  # Move to next ticket
-            logging.info("Fetching detail %d/%d: ticket %s", index, len(tickets), tid)  # Progress log
-            detail = OrgTicketManager._fetch_ticket_detail(org_id, tid)  # Get full ticket data
-            if detail:  # Only include tickets that returned data
-                all_details.append(DataProcessingUtils.flatten_dict(detail))  # Flatten and add
-            logging.debug("Fetched detail %d/%d", index, len(tickets))  # Progress debug log
-        return all_details  # All flattened records (may be empty)
-
-    @staticmethod
-    def _display_ticket_detail(ticket_data: dict) -> None:  # Display ticket detail.
-        """Format and display a ticket with its full comment history."""
-        print("\n  " + "=" * 60)  # Top separator bar
-        meta_fields = (  # (label, key, default) for the metadata block
-            ("Ticket", "subject", "(no subject)"),
-            ("ID    ", "id", "unknown"),
-            ("Status", "status", "unknown"),
-            ("Type  ", "type", "unknown"),
-            ("Created", "created_at", "unknown"),
-            ("Updated", "updated_at", "unknown"),
-        )
-        for label, key, default in meta_fields:  # Render header rows
-            print(f"  {label}: {ticket_data.get(key, default)}")  # Display ticket metadata row
-        print("  " + "-" * 60)  # Section separator
-        OrgTicketManager._render_comments_block(ticket_data.get("comments", []))  # Render comments
-
-    @staticmethod
-    def _render_comments_block(comments: list) -> None:
-        """Render the comments section: header + one block per comment + attachments."""
-        if not comments:  # No comments on this ticket
-            print("  No comments on this ticket.")  # Inform user
-            return  # Nothing else to render
-        print(f"  Comments ({len(comments)}):")  # Comment section header with count
-        for idx, comment in enumerate(comments, 1):  # Iterate each comment
-            author = comment.get("author", "unknown")  # Get comment author name
-            created = comment.get("created_at", "unknown")  # Get comment timestamp
-            text = comment.get("comment", "(no text)")  # Get comment body text
-            print(f"\n  [{idx}] {author} -- {created}")  # Comment header with author+date
-            print(f"      {text}")  # Comment body text indented
-            for att in comment.get("attachments", []) or []:  # Iterate attachments (may be empty)
-                print(f"      Attachment: {att.get('name', att.get('content_url', 'file'))}")
-
-        print("  " + "=" * 60)  # Bottom separator bar
+# OrgTicketManager moved to src/org/org_ticket_manager.py (1013 SC-001 Cat B position 46)
 
 
 # OrgAlarmEventExporter moved to src/export/org_alarm_event_exporter.py (1013 SC-001 position 18)
